Add technical notes
