+++ meta/11-1-1.md
@@ -62,5 +62,4 @@ source_geographical_coverage_1: Canada
 auto_progress_calculation: false
 progress_status: not_available
 ---
-<i>This indicator does not display a progress status as there is only one data point, therefore progress cannot be calculated.</i>
-
+<i>This indicator does not display a progress status as there is only one data point currently available.</i>
