+++ meta/11-2-1.md
@@ -60,7 +60,5 @@ source_geographical_coverage_1: Canada
 auto_progress_calculation: true
 progress_calculation_options:
 - direction: negative
-progress_status: not_available
+progress_status: substantial_progress
 ---
-<i>This indicator does not display a progress status as there is only one data point, therefore progress cannot be calculated.</i>
-
