+++ meta/11-5-1.md
@@ -83,4 +83,4 @@ progress_calculation_options:
   target_year: 2030
 progress_status: not_available
 ---
-<i>This indicator does not display a progress status as there is only one data point, therefore progress cannot be calculated.</i>
+<i>This indicator does not display a progress status as there is only one data point currently available.</i>
