+++ meta/12-2-1.md
@@ -71,5 +71,4 @@ progress_calculation_options:
 - direction: positive
 progress_status: not_available
 ---
-<i>This indicator does not display a progress status as there is only one data point, therefore progress cannot be calculated.</i>
-
+<i>This indicator does not display a progress status as there is only one data point currently available.</i>
