+++ meta/13-1-1.md
@@ -66,7 +66,8 @@ graph_annotations:
   value: 443.23
   preset: target_line
 
-data_footnote: The 2030 target correspond to the target for Canada, National Inventory Total only
+data_footnote: The 2030 target correspond to the target for Canada, National Inventory
+  Total only
 
 data_start_values:
 - field: Geography
