+++ meta/13-2-1.md
@@ -62,3 +62,4 @@ source_geographical_coverage_1: Canada
 auto_progress_calculation: false
 progress_status: not_available
 ---
+<i>Natural disasters are an indicator of climate change's impact over a long period of time; it is trivial to calculate progress over the current relative short period. Calculating progress over a short period with currently available data will result in an unclear measure of the effect of climate policies.</i>
