+++ meta/13-3-1.md
@@ -84,7 +84,7 @@ source_url_2: https://doi.org/10.25318/3410026101-eng
 source_organisation_2: Statistics Canada & Infrastructure Canada
 source_periodicity_2: Every 2 years
 source_geographical_coverage_2: Canada, provinces and territories
-auto_progress_calculation: true
+auto_progress_calculation: false
 progress_calculation_options:
 - direction: positive
 progress_status: substantial_progress
