+++ meta/15-1-1.md
@@ -68,6 +68,5 @@ progress_calculation_options:
 - direction: positive
 progress_status: not_available
 ---
-<i>This indicator does not display a progress status as there is only one data point, therefore progress cannot be calculated.</i>
-<br>
+<i>This indicator does not display a progress status as there is only one data point currently available.</i><br><br>
 This indicator corresponds to the Canadian Environmental Sustainability Indicators <a href="https://www.canada.ca/en/environment-climate-change/services/environmental-indicators/status-wild-species.html"> <em>Status of wild species</em></a>.
