+++ meta/16-3-1.md
@@ -26,7 +26,10 @@ computation_calculations: Not applicable
 
 comments_limitations: <p>During the production of each year's crime statistics, data
   from the previous year are revised to reflect any updates or changes that have been
-  received from the police services.</p><p>The total line "Total violent Criminal Code violations" aggregates the selecte types of crime reported in the chart above, as well as a few violation types not reported in the chart. See source table for all violation types.</p>
+  received from the police services.</p><p>The total line "Total violent Criminal
+  Code violations" aggregates the selected types of crime reported in the chart above,
+  as well as a few violation types not reported in the chart. See source table for
+  all violation types.</p>
 
 
 # Graph
@@ -54,5 +57,5 @@ source_organisation_1: Statistics Canada
 source_periodicity_1: Annual
 source_geographical_coverage_1: Canada, provinces and territories
 auto_progress_calculation: false
-progress_status: significant_deterioration
---- 
+progress_status: not_available
+---
