+++ meta/3-10-1.md
@@ -52,7 +52,7 @@ source_url_1: https://doi.org/10.25318/1310039401-eng
 source_organisation_1: Statistics Canada
 source_periodicity_1: Annual
 source_geographical_coverage_1: Canada
-auto_progress_calculation: true
+auto_progress_calculation: false
 progress_calculation_options:
 - direction: negative
 progress_status: moderate_progress
