+++ meta/3-8-1.md
@@ -66,3 +66,4 @@ source_geographical_coverage_2: Canada
 auto_progress_calculation: false
 progress_status: not_available
 ---
+<i>This indicator does not display a progress status as there is only one data point available for the target age groups.</i>
