+++ meta/5-3-1.md
@@ -49,4 +49,4 @@ progress_calculation_options:
   target: 50
 progress_status: not_available
 ---
-<i>This indicator does not display a progress status as there is only one data point, therefore progress cannot be calculated.</i>
+<i>This indicator does not display a progress status as there is only one data point currently available.</i>
